refactor(nxcdb): move shared fdatabase functions to single file
